+++ tests/queue.test.ts
@@ -1,4 +1,3 @@
-import { default as IORedis } from 'ioredis';
 import {
   describe,
   beforeEach,
@@ -14,18 +13,19 @@ import * as sinon from 'sinon';
 import { FlowProducer, Job, Queue, Worker } from '../src/classes';
 import { delay, randomUUID, removeAllQueueData } from '../src/utils';
 import { version as currentPackageVersion } from '../src/version';
+import { createTestConnection } from './utils/connection-factory';
+import { IRedisClient } from '../src/interfaces';
 
 describe('queues', () => {
-  const redisHost = process.env.REDIS_HOST || 'localhost';
   const prefix = process.env.BULLMQ_TEST_PREFIX || 'bull';
   const sandbox = sinon.createSandbox();
 
   let queue: Queue;
   let queueName: string;
 
-  let connection: IORedis;
+  let connection: IRedisClient;
   beforeAll(async () => {
-    connection = new IORedis(redisHost, { maxRetriesPerRequest: null });
+    connection = createTestConnection();
   });
 
   beforeEach(async () => {
@@ -37,7 +37,7 @@ describe('queues', () => {
   afterEach(async () => {
     sandbox.restore();
     await queue.close();
-    await removeAllQueueData(new IORedis(redisHost), queueName);
+    await removeAllQueueData(createTestConnection(), queueName);
   });
 
   afterAll(async function () {
@@ -103,7 +103,7 @@ describe('queues', () => {
     const version = await queue.getVersion();
     expect(version).toBe(null);
     await queue.close();
-    await removeAllQueueData(new IORedis(redisHost), exQueueName);
+    await removeAllQueueData(createTestConnection(), exQueueName);
   });
 
   describe('.getMeta', () => {
@@ -392,7 +392,7 @@ describe('queues', () => {
             expect(parentWaitCount).toEqual(1);
             await parentQueue.close();
             await flow.close();
-            await removeAllQueueData(new IORedis(redisHost), parentQueueName);
+            await removeAllQueueData(createTestConnection(), parentQueueName);
           });
         });
 
@@ -440,7 +440,7 @@ describe('queues', () => {
             expect(parentWaitCount).toEqual(1);
             await parentQueue.close();
             await flow.close();
-            await removeAllQueueData(new IORedis(redisHost), parentQueueName);
+            await removeAllQueueData(createTestConnection(), parentQueueName);
           });
         });
       });